back-end-main/api/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
anonypy==0.1.7
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
crypto==1.4.1
cycler==0.11.0
defusedxml==0.7.1
entrypoints==0.4
Faker==14.1.1
fastjsonschema==2.16.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.37.1
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
inflect==6.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.15.0
kiwisolver==1.4.4
lxml==4.9.1
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
matplotlib==3.5.3
mistune==2.0.4
multidict==6.0.2
mysqlclient==2.1.1
Naked==0.1.31
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
pyasn1==0.4.8
pydantic==1.9.2
Pygments==2.13.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.28.1
rsa==4.9
schedule==1.1.0
scipy==1.9.1
shellescape==3.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlacodegen==2.3.0
SQLAlchemy==1.4.39
SQLAlchemy-Utils==0.38.3
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
webencodings==0.5.1
Werkzeug==2.2.2
yarl==1.8.1
